+++ SRC/Server/common/Headers/CG_Packet.h
@@ -573,7 +573,7 @@ typedef struct SPacketCGMyShop
 typedef struct SPacketCGRefine
 {
 	uint8_t	header;
-	uint8_t	pos;
+	uint16_t	pos;
 	uint8_t	type;
 #ifdef ENABLE_FEATURES_REFINE_SYSTEM
 	uint8_t		lLow;
@@ -586,7 +586,7 @@ typedef struct SPacketCGRefine
 typedef struct SPacketCGRequestRefineInfo
 {
 	uint8_t	header;
-	uint8_t	pos;
+	uint16_t	pos;
 } TPacketCGRequestRefineInfo;
 
 typedef struct SPacketCGChangeName
